requirements.txt
nose
docker-py
requests